+++ src/js/background/model.js
@@ -2,8 +2,11 @@
     "use strict";
 
     window.ModelHelper = function (b) {
-        let shareUserdata = null;
         let data = {};
+        let shareInfo = {
+            config: null,
+            activity: null
+        };
 
         /**
          *
@@ -11,9 +14,11 @@
          */
         this.init = () => {
             return new Promise((resolve) => {
-                chrome.storage.sync.get(["model", "shareUserdata"], (obj) => {
+                chrome.storage.sync.get(["model", "shareInfo"], (obj) => {
                     data = obj.model || {};
-                    shareUserdata = typeof obj.shareUserdata === "undefined" ? null : obj.shareUserdata;
+                    if (typeof obj.shareInfo === "object") {
+                        shareInfo = obj.shareInfo;
+                    }
 
                     if (typeof data.installationDate === "undefined") { // no date yet -> save a start date in storage
                         data.installationDate = +new Date();
@@ -30,10 +35,42 @@
             });
         };
 
-        this.shareUserdata = () => {
-            return shareUserdata;
+        /**
+         * Returns information about what the user allows to be tracked
+         *
+         * @returns {object}
+         */
+        this.getShareInfo = () => shareInfo;
+
+        /**
+         * Sets the information about what the users wants to be tracked
+         *
+         * @param {object} opts
+         * @returns {Promise}
+         */
+        this.setShareInfo = (opts) => {
+            return new Promise((resolve) => {
+                shareInfo = {
+                    config: opts.config || false,
+                    activity: opts.activity || false
+                };
+
+                chrome.storage.sync.set({
+                    shareInfo: shareInfo
+                }, () => {
+                    chrome.runtime.lastError; // do nothing specific with the error -> is thrown if too many save attempts are triggered
+                    resolve();
+                });
+            });
         };
 
+        /**
+         * Saves the given value under the given name
+         *
+         * @param {string} key
+         * @param {*} val
+         * @returns {Promise}
+         */
         this.setData = (key, val) => {
             return new Promise((resolve) => {
                 data[key] = val;
@@ -41,10 +78,21 @@
             });
         };
 
+        /**
+         * Returns the value to the given name
+         *
+         * @param {string} key
+         * @returns {*|null}
+         */
         this.getData = (key) => {
             return data[key] || null;
         };
 
+        /**
+         * Saves the data object into the synced storage
+         *
+         * @returns {Promise}
+         */
         let saveData = () => {
             return new Promise((resolve) => {
                 if (Object.getOwnPropertyNames(data).length > 0) {