+++ src/js/helper/toggle.js
@@ -88,11 +88,12 @@
                 ext.elements.iframe.addClass(ext.opts.classes.page.visible);
                 ext.addReloadMask();
             } else {
-                ext.helper.model.call("shareUserdataMask").then((opts) => { // check whether to show the share userdata mask or not
+                ext.helper.model.call("shareInfoMask").then((opts) => { // check whether to show the share userdata mask or not
+                    opts.showMask=true;
                     if (opts && opts.showMask) {
-                        ext.addShareUserdataMask();
+                        ext.addShareInfoMask();
                     } else {
-                        ext.elements.sidebar.find("#" + ext.opts.ids.sidebar.shareUserdata).remove();
+                        ext.elements.sidebar.find("#" + ext.opts.ids.sidebar.shareInfo).remove();
                     }
                 });